requirements.txt
Flask==1.1.2
joblib==0.16.0
matplotlib==3.3.1
nltk==3.5
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
plotly==4.9.0
scikit-learn==0.23.2
scipy==1.5.2
torch==1.6.0+cpu
torchvision==0.7.0+cpu
waitress==1.4.4
pyro-ppl == 1.4.0